+++ Package.swift
@@ -16,6 +16,9 @@ let package = Package(
             name: "FirestoreAdminServer",
             targets: ["FirestoreAdminServer"]),
         .library(
+            name: "FirestoreEmbedded",
+            targets: ["FirestoreEmbedded"]),
+        .library(
             name: "FirestoreMongoCore",
             targets: ["FirestoreMongoCore"]),
     ],
@@ -28,6 +31,8 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-log.git", from: "1.14.0")
     ],
     targets: [
+        .target(
+            name: "FirestoreEmbedded"),
         .target(
             name: "FirestoreCore"),
         .target(
@@ -188,6 +193,7 @@ let package = Package(
                 "FirestoreAdminGRPCBootstrap",
                 "FirestoreAuthCore",
                 "FirestoreAuth",
+                "FirestoreEmbedded",
                 "FirestoreCodable",
                 "FirestoreGeoQuery",
                 "FirestoreMongoCore",