+++ Sources/FirestoreEmbedded/FirestoreEmbeddedError.swift
@@ -1,0 +1,4 @@
+public enum FirestoreEmbeddedError: Error, Equatable, Sendable {
+    case invalidPath(String)
+    case invalidValue(String)
+}